update: update filed type and moved retail
